requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
python-dotenv==1.0.0
openai==1.3.0
PyPDF2==3.0.1
langchain==0.0.350
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.5.2 